+++ include/sqlpp23/postgresql/prepared_statement.h
@@ -34,6 +34,7 @@
 
 #include <sqlpp23/core/chrono.h>
 #include <sqlpp23/core/debug_logger.h>
+#include <sqlpp23/core/to_sql_string.h>
 #include <sqlpp23/postgresql/database/connection_handle.h>
 #include <sqlpp23/postgresql/database/serializer_context.h>
 #include <sqlpp23/postgresql/pg_result.h>
@@ -129,9 +130,9 @@ class prepared_statement_t {
 
     _stmt_null_parameters[index] = false;
     if (value) {
-      _stmt_parameters[index] = "TRUE";
+      _stmt_parameters[index] = "t";
     } else {
-      _stmt_parameters[index] = "FALSE";
+      _stmt_parameters[index] = "f";
     }
   }
 
@@ -144,6 +145,7 @@ class prepared_statement_t {
 
     _stmt_null_parameters[index] = false;
     context_t context{nullptr};
+    using sqlpp::to_sql_string;
     _stmt_parameters[index] = to_sql_string(context, value);
   }