+++ tests/core/constraints/clause/union.cpp
@@ -107,7 +107,27 @@ int main() {
     CANNOT_CALL_ANY_UNION_WITH(lhs, bad_custom_rhs);
   }
 
-  // UNION requires statements with same result row
+  // UNION requires statements with same result row (single column)
+  {
+    auto s_foo_int = select(foo.id).from(foo);
+    auto s_foo_int_n = select(foo.intN).from(foo);
+    auto s_value_id = select(sqlpp::value(7).as(foo.id)).from(foo);
+    auto s_value_oid =
+        select(sqlpp::value(7).as(something)).from(foo);
+    // Different value type
+    static_assert(
+        not std::is_same<sqlpp::data_type_of_t<decltype(foo.id)>,
+                         sqlpp::data_type_of_t<decltype(foo.intN)>>::value,
+        "");
+    CANNOT_CALL_ANY_UNION_WITH(s_foo_int, s_foo_int_n);
+    // Different name
+    CANNOT_CALL_ANY_UNION_WITH(s_value_id, s_value_oid);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo_int, dynamic(maybe, s_foo_int_n));
+    // Different name
+    CANNOT_CALL_ANY_UNION_WITH(s_value_id, dynamic(maybe, s_value_oid));
+  }
+
+  // UNION requires statements with same result row (more than one column)
   {
     auto s_foo_int = select(foo.textNnD, foo.id).from(foo);
     auto s_foo_int_n = select(foo.textNnD, foo.intN).from(foo);
@@ -127,28 +147,124 @@ int main() {
     CANNOT_CALL_ANY_UNION_WITH(s_value_id, dynamic(maybe, s_value_oid));
   }
 
-  // UNION arguments must not contain order_by, because the resulting SQL would
-  // be invalid if
-  //   - the order_by were in the left query
-  //   - the order_by were in the right query and the arguments were using
-  //     different tables (columns are serialized with table-qualification).
+  // UNION arguments must not contain for_update.
+  {
+    auto s_foo = select(foo.id).from(foo);
+    auto s_bar = select(bar.id).from(bar);
+
+    // Do not allow for_update in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.for_update(), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.for_update(), s_bar);
+
+    // Do not allow for_update in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.for_update());
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.for_update());
+
+    // Do not allow for_update in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.for_update(),
+                               s_bar.for_update());
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.for_update(),
+                               s_bar.for_update());
+  }
+
+  // UNION arguments must not contain order_by.
+  {
+    auto s_foo = select(foo.id).from(foo);
+    auto s_bar = select(bar.id).from(bar);
+
+    // Do not allow order_by in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_bar);
+
+    // Do not allow order_by in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.order_by(foo.id.asc()));
+
+    // Do not allow order_by in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+  }
+
+  // UNION arguments must not contain union_order_by.
+  {
+    auto s_foo = select(foo.id).from(foo).union_all(select(foo.id).from(foo));
+    auto s_bar = select(bar.id).from(bar);
+
+    // Do not allow order_by in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_bar);
+
+    // Do not allow order_by in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.order_by(foo.id.asc()));
+
+    // Do not allow order_by in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+  }
+
+  // UNION arguments must not contain order_by.
+  {
+    auto s_foo = select(foo.id).from(foo);
+    auto s_bar = select(bar.id).from(bar);
+
+    // Do not allow order_by in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.asc()), s_bar);
+
+    // Do not allow order_by in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.order_by(foo.id.asc()));
+
+    // Do not allow order_by in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.asc()),
+                               s_bar.order_by(foo.id.asc()));
+  }
+
+  // UNION arguments must not contain limit.
+  {
+    auto s_foo = select(foo.id).from(foo);
+    auto s_bar = select(bar.id).from(bar);
+
+    // Do not allow limit in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.limit(7), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.limit(7), s_bar);
+
+    // Do not allow limit in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.limit(7));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.limit(7));
+
+    // Do not allow limit in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.limit(7),
+                               s_bar.limit(7));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.limit(7),
+                               s_bar.limit(7));
+  }
+
+  // UNION arguments must not contain offset.
   {
     auto s_foo = select(foo.id).from(foo);
     auto s_bar = select(bar.id).from(foo);
 
-    // Do not allow order_by in the LHS expression.
-    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.desc()), s_foo);
-    CANNOT_CALL_ANY_UNION_WITH(s_foo.order_by(foo.id.desc()), s_bar);
+    // Do not allow offset in the LHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.offset(42), s_foo);
+    CANNOT_CALL_ANY_UNION_WITH(s_foo.offset(42), s_bar);
 
-    // Do not allow order_by in the RHS expression.
-    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.order_by(foo.id.desc()));
-    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.order_by(foo.id.desc()));
+    // Do not allow offset in the RHS expression.
+    CANNOT_CALL_ANY_UNION_WITH(s_foo, s_bar.offset(42));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar, s_bar.offset(42));
 
-    // Do not allow order_by in both expressions.
-    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.desc()),
-                               s_bar.order_by(foo.id.desc()));
-    CANNOT_CALL_ANY_UNION_WITH(s_bar.order_by(foo.id.desc()),
-                               s_bar.order_by(bar.id.desc()));
+    // Do not allow offset in both expressions.
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.offset(42),
+                               s_bar.offset(42));
+    CANNOT_CALL_ANY_UNION_WITH(s_bar.offset(42),
+                               s_bar.offset(42));
   }
 
   // UNION requires preparable statements
@@ -184,5 +300,4 @@ int main() {
             sqlpp::statement_prepare_check_t<U>,
             sqlpp::assert_no_unknown_tables_in_selected_columns_t>::value);
   }
-
 }